+++ src/main/scala/com/gu/hstschecker/reports/Report.scala
@@ -5,17 +5,19 @@ import java.io.PrintStream
 import fansi.Str
 
 case class Report(lines: List[Str]) {
-  def +(line: Str) = this.copy(this.lines :+ line)
+  def +(line: Str): Report = this.copy(this.lines :+ line)
 }
 
 object Report {
+  case class FailureReason(message: String) extends AnyVal
+
   def outputReport(report: Report, stream: PrintStream, ansi: Boolean = true): Unit = {
     report.lines.foreach { line =>
       stream.println(if (ansi) line else line.plainText)
     }
   }
 
-  def printReports(results: List[Either[Report, Option[Report]]]) = {
+  def printReports(results: List[Either[Report, Option[Report]]]): Unit = {
     val (errorReports, resultReports) = results.foldLeft[(List[Report], List[Report])](Nil, Nil) { case ((errAcc, resAcc), either) =>
       either.fold(
         err => (errAcc :+ err, resAcc),
